+++ src/Patch/TerminalPatch.cs
@@ -67,27 +67,32 @@ internal class TerminalPatch
     {
         if (__result is null) return null;
 
-        var terminalInput = __instance.screenText.text.Substring(__instance.screenText.text.Length - __instance.textAdded).ToLower();
+        var terminalInput = __instance.screenText.text[^__instance.textAdded..].ToLower().Trim();
         var inputArray = terminalInput.Split(" ").Where(x => !x.IsNullOrWhiteSpace()).ToList();
-        var inputCommand = inputArray[0];
-        var inputCommandArgs = inputArray.Count > 1 ? inputArray[1] : null;
-        //_logger.LogDebug($"TryReturnSpecialNodes: {__result.name} | input: {terminalInput}");
+        var inputCommand = terminalInput.Contains(" ") ? terminalInput[..terminalInput.IndexOf(' ')] : terminalInput;
+        var inputCommandArgs = terminalInput.Contains(" ") ? terminalInput[(terminalInput.IndexOf(' ') + 1)..] : null;
+        // var inputCommand = inputArray[0];
+        // var inputCommandArgs = string.Concat(inputArray.Skip(1));//inputArray.Count > 1 ? inputArray[1] : null;
+        //Logger.LogDebug($"TryReturnSpecialNodes: {__result.name} | input: {terminalInput}");
 
         // Try to find the matching primary command first
-        var filteredCommands = AdvancedTerminal.Commands.Where(x => x.CommandText == inputCommand).ToList();
+        var filteredCommands = AdvancedTerminal.Commands
+            .Where(x => inputCommand.Length < 3 ? x.CommandText == inputCommand : x.CommandText.StartsWith(inputCommand))
+            .ToList();
+
         if (!filteredCommands.Any())
         {
-            //_logger.LogDebug($" > No commands found matching '{inputCommand}' with args '{inputCommandArgs}'");
+            //Logger.LogDebug($" > No commands found matching '{inputCommand}' with args '{inputCommandArgs}'");
             return __result;
         }
 
         if (filteredCommands.Count > 1)
         {
-            Logger.LogError($" > Found multiple commands! HOW? Only using first one. Found: {filteredCommands.Select(x => x.CommandText).Aggregate((first, second) => $"{first}, {second}")}");
+            //Logger.LogError($" > Found multiple commands! HOW? Only using first one. Found: {filteredCommands.Select(x => x.CommandText).Aggregate((first, second) => $"{first}, {second}")}");
         }
 
         var advancedCommand = filteredCommands.First();
-        //_logger.LogDebug($" > command: {advancedCommand.CommandText} | event: /{__result.terminalEvent}/");
+        //Logger.LogDebug($" > command: {advancedCommand.CommandText} | event: /{__result.terminalEvent ?? "empty"}/");
 
         if (advancedCommand.IsSimpleCommand)
         {
@@ -101,30 +106,30 @@ internal class TerminalPatch
             if (resNode != null) return resNode;
         }
 
-        //_logger.LogDebug($"TryReturnSpecialNodes.end");
+        //Logger.LogDebug($"TryReturnSpecialNodes.end");
         return __result;
     }
 
     private static TerminalNode ExecuteSimpleCommand(TerminalCommandBuilder command)
     {
-        //_logger.LogDebug(" > executing SimpleCommand");
-        //_logger.LogDebug("  > checking conditions");
+        //Logger.LogDebug(" > executing SimpleCommand");
+        //Logger.LogDebug("  > checking conditions");
         foreach (var (node, condition) in command.SpecialNodes)
         {
-            //_logger.LogDebug($"   > condition: {node.name}");
+            //Logger.LogDebug($"   > condition: {node.name}");
 
             if (!condition()) continue;
 
-            //_logger.LogDebug($"    > FAILED");
+            //Logger.LogDebug($"    > FAILED");
             return node;
         }
 
-        return null;
+        return command.Node;
     }
 
     private static TerminalNode ExecuteComplexCommand(TerminalCommandBuilder command, TerminalNode __result, string inputCommand, string inputCommandArgs)
     {
-        //_logger.LogDebug(" > executing ComplexCommand");
+        //Logger.LogDebug(" > executing ComplexCommand");
 
         // Try to find a matching non-variable (input) command
         var subCommand = command.SubCommands.FirstOrDefault(subCmd => !subCmd.IsVariableCommand && subCmd.Name == inputCommandArgs);
@@ -143,7 +148,7 @@ internal class TerminalPatch
             {
                 if (!subCmd.IsVariableCommand) return false;
 
-                //_logger.LogDebug($"  > INPUT: '{inputCommand}' | '{inputCommandArgs}' | REGEX: {subCmd.VariableRegexMatchPattern}");
+                //Logger.LogDebug($"  > INPUT: '{inputCommand}' | '{inputCommandArgs}' | REGEX: {subCmd.VariableRegexMatchPattern}");
                 var regex = new Regex(subCmd.VariableRegexMatchPattern);
                 var match = regex.Match(inputCommandArgs);
                 return match.Success;
@@ -154,25 +159,25 @@ internal class TerminalPatch
         // If nothing was found, just return the main commands node
         if (subCommand is null)
         {
-            //_logger.LogDebug($"  > No matching sub command found for input: '{inputCommand}' | '{inputCommandArgs}'");
+            //Logger.LogDebug($"  > No matching sub command found for input: '{inputCommand}' | '{inputCommandArgs}'");
             return command.Node;
         }
 
-        //_logger.LogDebug("  > checking special conditions");
+        //Logger.LogDebug("  > checking special conditions");
         foreach (var conditionString in subCommand.Conditions)
         {
             var specialCondition = command.SpecialNodes.FirstOrDefault(x => x.node.name == conditionString);
             if (specialCondition == default)
             {
-                Logger.LogError($"> SubCommand {subCommand.Name} has special criteria for '{conditionString}' but it is not found as part of the commands' special conditions list.");
+                //Logger.LogError($"> SubCommand {subCommand.Name} has special criteria for '{conditionString}' but it is not found as part of the commands' special conditions list.");
                 break;
             }
 
-            //_logger.LogDebug($"   > {conditionString} = {specialCondition.condition()}");
+            //Logger.LogDebug($"   > {conditionString} = {specialCondition.condition()}");
 
             if (specialCondition.condition()) continue;
 
-            //_logger.LogDebug($"   > FAILED: {specialCondition.node.name}");
+            //Logger.LogDebug($"   > FAILED: {specialCondition.node.name}");
             return specialCondition.node;
         }
 
@@ -184,18 +189,18 @@ internal class TerminalPatch
     [HarmonyPatch(nameof(Terminal.RunTerminalEvents))]
     public static void RunTerminalEventsPatch(TerminalNode node)
     {
-        //_logger.LogDebug($"RunTerminalEvents: node: {node?.ToString() ?? "null"} | terminalEvent: {node?.terminalEvent ?? "empty"}");
+        //Logger.LogDebug($"RunTerminalEvents: node: {node?.ToString() ?? "null"} | terminalEvent: {node?.terminalEvent ?? "empty"}");
 
         if (node.terminalEvent.IsNullOrWhiteSpace()) return;
 
         foreach (var command in AdvancedTerminal.Commands)
         {
-            //_logger.LogDebug($"  > Checking {command.node.name}");
+            //Logger.LogDebug($"  > Checking {command.Node.name}");
 
             if (command.IsSimpleCommand && command.ActionEvent == node.terminalEvent)
             {
-                //_logger.LogDebug("  > IsSimpleCommand, SKIP??");
-                //_logger.LogDebug($" > EXEC simple command: {command.node.name}");
+                //Logger.LogDebug("  > IsSimpleCommand, SKIP??");
+                //Logger.LogDebug($" > EXEC simple command: {command.Node.name}");
                 var res = command.Action();
                 node.displayText = res + AdvancedTerminal.EndOfMessage;
 
@@ -205,7 +210,7 @@ internal class TerminalPatch
             var match = command.SubCommands.FirstOrDefault(x => x.ActionEvent == node.terminalEvent);
             if (match == default) continue;
 
-            //_logger.LogDebug($"  > Found {match.Node.name}");
+            //Logger.LogDebug($"  > Found {match.Node.name}");
             match.Action();
 
             break;