+++ src/TerminalCommands/SellCommands.cs
@@ -22,37 +22,64 @@ internal class SellCommands
         if (!Plugin.Instance.PluginConfig.TerminalSellCommandsEnabled) return null;
 
         return new TerminalCommandBuilder("sell")
-            .WithDescription("> SELL [ALL|QUOTA|TARGET|2h|<AMOUNT>|<ITEM>]\nTo sell items on the ship. Will not sell ignored items from config.")
+            .WithDescription("> sell [ALL|QUOTA|TARGET|2h|<AMOUNT>|<ITEM>]\nTo sell items on the ship. Will not sell ignored items from config except when selling a specific item.")
             .WithText("Please enter [ALL|QUOTA|TARGET|2h|<AMOUNT>|<ITEM>]")
             .WithSubCommand(CreateAllSubCommand())
             .WithSubCommand(CreateQuotaSubCommand())
             .WithSubCommand(CreateTargetSubCommand())
-            .WithSubCommand(CreateAmountSubCommand())
             .WithSubCommand(Create2HandedSubCommand())
+            .WithSubCommand(CreateAmountSubCommand())
             .WithSubCommand(CreateItemSubCommand())
-            .AddTextReplacement("[sellScrapFor]", () => _sellScrapFor.ToString())
-            .AddTextReplacement("[sellScrapTarget]", () => _sellScrapActualTarget.ToString())
-            .AddTextReplacement("[numScrapSold]", () => _recommendedScraps.Count.ToString())
-            .AddTextReplacement("[shipTotalScrapCount]", () => ScrapUtils.GetShipSellableScrapCount().ToString())
-            .AddTextReplacement("[shipTotalScrapValue]", () => ScrapUtils.GetShipTotalSellableScrapValue().ToString())
-            .AddTextReplacement("[sellScrapActualTotal]", () => ScrapUtils.SumScrapListSellValue(_recommendedScraps).ToString())
-            .AddTextReplacement("[companyBuyItemsCombo]", () => _recommendedScraps?.Select(x => $"{x.itemProperties.name}: {x.ActualSellValue()}").Aggregate((first, next) => $"{first}\n{next}"))
+            .AddTextReplacement("[sellScrapFor]", () => $"${_sellScrapFor}")
+            .AddTextReplacement("[sellScrapTarget]", () => $"${_sellScrapActualTarget}")
+            .AddTextReplacement("[numScrapSold]", () => _recommendedScraps.Count)
+            .AddTextReplacement("[sellScrapActualTotal]", () =>
+            {
+                if (GameUtils.IsCompanyBuyingAtFullRate())
+                {
+                    return $"${_recommendedScraps.ScrapValueOfCollection()}";
+                }
+
+                return $"${_recommendedScraps.ActualScrapValueOfCollection()} (${_recommendedScraps.ScrapValueOfCollection()})";
+            })
+            .AddTextReplacement("[companyBuyItemsCombo]", GenerateBuyItemsComboText)
             .WithCondition("landedAtCompany", "ERROR: Usage of this feature is only permitted within Company bounds\n\nPlease land at 71-Gordion and repeat command.", GameUtils.IsLandedOnCompany)
-            .WithCondition("hasScrapItems", "Bruh, you don't even have any items.", () => ScrapUtils.GetShipSellableScrapCount() > 0)
-            .WithCondition("notEnoughScrap", "Not enough scrap to meet [sellScrapFor] credits.\nTotal value: [shipTotalScrapValue].", () => _sellScrapFor < ScrapUtils.GetShipTotalSellableScrapValue())
-            .WithCondition("quotaAlreadyMet", "Quota already met.", () => TimeOfDay.Instance.profitQuota - TimeOfDay.Instance.quotaFulfilled > 0)
+            .WithCondition("hasScrapItems", "Bruh, you don't even have any items.", () => ScrapUtils.GetAllScrapInShip().Count > 0)
+            .WithCondition("notEnoughScrap", "Not enough scrap to meet [sellScrapFor] credits.\nTotal value: [sellScrapActualTotal].", () => _sellScrapFor < ScrapUtils.GetShipTotalSellableScrapValue())
+            .WithCondition("quotaAlreadyMet", "Quota has already been met.", () => TimeOfDay.Instance.profitQuota - TimeOfDay.Instance.quotaFulfilled > 0)
             .WithCondition("hasMatchingScrapItems", "No matching items found for input.", () => _recommendedScraps.Count > 0)
-            .WithCondition("targetCommandDisabled", "Target command has been disabled", () => Plugin.Instance.PluginConfig.TerminalTargetCommandsEnabled);
+            .WithCondition("targetCommandDisabled", "Target command has been disabled", () => Plugin.Instance.PluginConfig.TerminalTargetCommandsEnabled)
+            .WithCondition("targetAlreadyMet", "Target has already been met.", () => InfoMonitor.Instance.CalculatedNeededToReachTarget > 0);
+
+        static string GenerateBuyItemsComboText()
+        {
+            if (_recommendedScraps is null || _recommendedScraps.Count == 0) return "No items";
+
+            if (GameUtils.IsCompanyBuyingAtFullRate())
+            {
+                return _recommendedScraps
+                    .Select(x => $"{x.itemProperties.name}: {x.scrapValue}")
+                    .Aggregate((first, next) => $"{first}\n{next}");
+            }
+
+            return _recommendedScraps
+                .Select(x => $"{x.itemProperties.name}: ${x.ActualSellValue()} (${x.scrapValue})")
+                .Aggregate((first, next) => $"{first}\n{next}");
+        }
     }
 
     private static TerminalSubCommandBuilder CreateAllSubCommand()
     {
         return new TerminalSubCommandBuilder("all")
             .WithDescription("Sell ALL items on the ship.")
-            .WithMessage("[companyBuyingRateWarning]Requesting to sell ALL scrap ([shipTotalScrapCount]) for $[shipTotalScrapValue] credits.")
-            .EnableConfirmDeny(confirmMessage: "Transaction complete. Sold [shipTotalScrapCount] scrap for $[shipTotalScrapValue] credits.\n\nThe company is not responsible for any calculation errors.")
+            .WithMessage("[companyBuyingRateWarning]Requesting to sell ALL scrap ([numScrapSold]) for [sellScrapActualTotal] credits.")
+            .EnableConfirmDeny(confirmMessage: "Transaction complete. Sold [numScrapSold] scrap for [sellScrapActualTotal] credits.\n\nThe company is not responsible for any calculation errors.")
             .WithConditions("landedAtCompany", "hasScrapItems")
-            .WithPreAction(() => _sellScrapFor = ScrapUtils.GetShipTotalSellableScrapValue())
+            .WithPreAction(() =>
+            {
+                _recommendedScraps = ScrapUtils.GetAllSellableScrapInShip();
+                _sellScrapFor = _recommendedScraps.ActualScrapValueOfCollection();
+            })
             .WithAction(() =>
             {
                 TargetManager.SellAllScrap();
@@ -63,8 +90,8 @@ internal class SellCommands
     {
         return new TerminalSubCommandBuilder("quota")
             .WithDescription("Sell only to meet the current quota.")
-            .WithMessage("[companyBuyingRateWarning]Requesting to sell scrap as close to current quota ($[sellScrapFor] credits) as possible...\nThe Company wants the follow items for a total of [sellScrapActualTotal]:\n[companyBuyItemsCombo]")
-            .EnableConfirmDeny(confirmMessage: "Transaction complete. Sold [shipTotalScrapCount] scrap for $[shipTotalScrapValue] credits.\n\nThe company is not responsible for any calculation errors.")
+            .WithMessage("[companyBuyingRateWarning]Requesting to sell scrap as close to current quota ([sellScrapFor] credits) as possible...\nThe Company wants the follow items for a total of [sellScrapActualTotal]:\n[companyBuyItemsCombo]")
+            .EnableConfirmDeny(confirmMessage: "Transaction complete. Sold [numScrapSold] scrap for [sellScrapActualTotal] credits.\n\nThe company is not responsible for any calculation errors.")
             .WithConditions("landedAtCompany", "hasScrapItems", "notEnoughScrap", "quotaAlreadyMet")
             .WithPreAction(() =>
             {
@@ -81,8 +108,8 @@ internal class SellCommands
     {
         return new TerminalSubCommandBuilder("target")
             .WithDescription("Sell only to meet the current set target. Will sell quota if target is below target.")
-            .WithMessage("[companyBuyingRateWarning]Requesting to sell scrap as close to current target ($[sellScrapTarget], needing $[sellScrapFor]) as possible...\nThe Company wants the follow items for a total of [sellScrapActualTotal]:\n[companyBuyItemsCombo]")
-            .EnableConfirmDeny(confirmMessage: "Transaction complete. Sold [shipTotalScrapCount] scrap for $[shipTotalScrapValue] credits.\n\nThe company is not responsible for any calculation errors.")
+            .WithMessage("[companyBuyingRateWarning]Requesting to sell scrap as close to current target ($[sellScrapTarget], needing [sellScrapFor]) as possible...\nThe Company wants the follow items for a total of [sellScrapActualTotal]:\n[companyBuyItemsCombo]")
+            .EnableConfirmDeny(confirmMessage: "Transaction complete. Sold [numScrapSold] scrap for [sellScrapActualTotal] credits.\n\nThe company is not responsible for any calculation errors.")
             .WithConditions("targetCommandDisabled", "landedAtCompany", "hasScrapItems", "notEnoughScrap", "targetAlreadyMet")
             .WithPreAction(() =>
             {
@@ -103,7 +130,7 @@ internal class SellCommands
     {
         return new TerminalSubCommandBuilder("<amount>")
             .WithDescription("Sell as close as possible to input amount")
-            .WithMessage("[companyBuyingRateWarning]Requesting to sell scrap as close to $[sellScrapFor] as possible...\n\nThe Company wants the follow items for a total of $[sellScrapActualTotal]:\n[companyBuyItemsCombo]")
+            .WithMessage("[companyBuyingRateWarning]Requesting to sell scrap as close to [sellScrapFor] as possible...\n\nThe Company wants the follow items for a total of $[sellScrapActualTotal]:\n[companyBuyItemsCombo]")
             .EnableConfirmDeny(confirmMessage: "Sold [numScrapSold] scrap for $[sellScrapActualTotal].\n\nThe company is not responsible for any calculation errors.")
             .WithConditions("landedAtCompany", "hasScrapItems", "notEnoughScrap")
             .WithInputMatch(@"^(\d+)$")
@@ -118,10 +145,8 @@ internal class SellCommands
                     .ThenByDescending(x => x.scrapValue)
                     .ToList();
 
-                // Nothing found, return notEnoughScrapNode
                 if (_recommendedScraps.Count == 0) return false;
 
-                // A combination has been found, return info with confirm/deny node
                 return true;
             })
             .WithAction(() =>
@@ -134,8 +159,8 @@ internal class SellCommands
     {
         return new TerminalSubCommandBuilder("2h")
             .WithDescription("Sell all 2handed items")
-            .WithMessage("[companyBuyingRateWarning]Requesting to sell all two-handed scrap.\n\nThe Company wants the follow items for a total of $[sellScrapActualTotal]:\n[companyBuyItemsCombo]")
-            .EnableConfirmDeny(confirmMessage: "Sold [numScrapSold] scrap for $[sellScrapActualTotal].\n\nThe company is not responsible for any calculation errors.")
+            .WithMessage("[companyBuyingRateWarning]Requesting to sell all two-handed scrap.\n\nThe Company wants the follow items for a total of [sellScrapActualTotal]:\n[companyBuyItemsCombo]")
+            .EnableConfirmDeny(confirmMessage: "Transaction complete. Sold [numScrapSold] scrap for [sellScrapActualTotal] credits.\n\nThe company is not responsible for any calculation errors.")
             .WithConditions("landedAtCompany", "hasScrapItems", "notEnoughScrap")
             .WithPreAction(() =>
             {
@@ -162,15 +187,13 @@ Examples:
 > sell shot>50   - sells shotguns with a value of more than 50
 > sell lamp<=100 - lamps less than or equal to 100"
             )
-            .WithMessage("[companyBuyingRateWarning]Requesting to sell specified items.\n\nThe Company wants the follow items for a total of $[sellScrapActualTotal]:\n[companyBuyItemsCombo]")
-            .EnableConfirmDeny(confirmMessage: "Sold [numScrapSold] scrap for $[sellScrapActualTotal].\n\nThe company is not responsible for any calculation errors.")
+            .WithMessage("[companyBuyingRateWarning]Requesting to sell all matching items.\n\nThe Company wants the follow items for a total of [sellScrapActualTotal]:\n[companyBuyItemsCombo]")
+            .EnableConfirmDeny(confirmMessage: "Transaction complete. Sold [numScrapSold] scrap for [sellScrapActualTotal] credits.\n\nThe company is not responsible for any calculation errors.")
             .WithConditions("landedAtCompany", "hasScrapItems", "hasMatchingScrapItems")
             .WithInputMatch("^(.+)$")
             .WithPreAction(input =>
             {
                 input = RemoveWhitespaceRegex.Replace(input, "");
-
-                Plugin.Instance.ACLogger.LogDebug($"> '{input}'");
 
                 var reg = new Regex(@"^(\w+)(>|<|>=|<=|=)(\d+)$");
                 var match = reg.Match(input);
@@ -184,10 +207,13 @@ Examples:
                 else
                 {
                     // conditional selling
+#if DEBUG
                     foreach (Group group in match.Groups)
                     {
                         Plugin.Instance.ACLogger.LogDebug($"> '{group.Value}'");
                     }
+#endif
+
                     var itemName = match.Groups[1].Value;
                     var greaterThan = match.Groups[2].Value == ">";
                     var smallerThan = match.Groups[2].Value == "<";
@@ -198,13 +224,15 @@ Examples:
                     _recommendedScraps = ScrapUtils.GetAllScrapInShip()
                         .Where(x => x.itemProperties.name.ToLower().Contains(itemName))
                         .Where(x =>
-                            greaterThan ? x.scrapValue > amount :
-                            smallerThan ? x.scrapValue < amount :
-                            greaterThanEqualTo ? x.scrapValue >= amount :
-                            smallerThanEqualTo ? x.scrapValue <= amount :
-                            x.scrapValue == amount
-                        )
-                        .OrderByDescending(x => x.scrapValue)
+                        {
+                            var scrapValue = x.ActualSellValue();
+                            return greaterThan ? scrapValue > amount :
+                                smallerThan ? scrapValue < amount :
+                                greaterThanEqualTo ? scrapValue >= amount :
+                                smallerThanEqualTo ? scrapValue <= amount :
+                                scrapValue == amount;
+                        })
+                        .OrderByDescending(x => x.ActualSellValue())
                         .ToList();
                 }